pyproject.toml
[project]
name = "qa-guru-belevtseva-hw-19"
version = "0.1.0"
description = ""
authors = [
    { name = "Darya Belevtseva", email = "dascha067@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.0.0)",
    "selenium (>=4.15.0)",
    "selene (>=2.0.0rc9)",
    "allure-pytest (>=2.13.0)",
    "appium-python-client (>=4.0.0)",
    "requests (>=2.32.0)",
    "python-dotenv (>=1.0.1)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
